Fix some trivial sign-compare compiler warnings
Acked-by: Gert Doering <gert@greenie.muc.de>
Message-Id: <20250114165206.13187-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg30455.html

+++ src/openvpn/options.c
@@ -1448,7 +1448,7 @@ foreign_options_copy_dns(struct options *o, struct env_set *es)
 
     while (server)
     {
-        for (int i = 0; i < server->addr_count; ++i)
+        for (size_t i = 0; i < server->addr_count; ++i)
         {
             if (server->addr[i].family == AF_INET)
             {
